// Server/src/auth/dto/register.dto.ts
import { z } from 'zod';

export const RegisterSchema = z.object({
  email: z.string().email('Invalid email address'),
  password: z.string().min(6, 'Password must be at least 6 characters long'),
  name: z.string().min(2, 'Name must be at least 2 characters long').optional(),
});

export type RegisterDto = z.infer<typeof RegisterSchema>;